Added SimpleCellImplementingLayoutAttributes
Now we can see how it handles a cell that only implements
preferredLayoutAttributesFittingAttributes. I learned that if we don’t
cache the size we return it will keep updating the frame, calling
preferredLayoutAttributes indefinitely and eventually crash. So what we
need is to actually cache it and return a size only once